[FLINK-22737][table-common] Move specific type strategies out of TypeStrategies
